Reverting the generic LinkInterceptor service change
