--- AngularApp/projects/app-service-diagnostics/src/app/resources/web-hosting-environments/services/ase-link-interceptor.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { AseLinkInterceptorService } from './ase-link-interceptor.service';
-
-describe('AseLinkInterceptorService', () => {
-  let service: AseLinkInterceptorService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(AseLinkInterceptorService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
