--- AngularApp/projects/app-service-diagnostics/src/app/resources/web-hosting-environments/services/ase-link-interceptor.service.ts
@@ -1,12 +1,0 @@
-import { Injectable } from '@angular/core';
-import { LinkInterceptorService } from '../../../shared-v2/services/link-interceptor.service';
-import { DiagnosticService } from 'diagnostic-data';
-import { AseCategoryService } from './ase-category.service';
-
-@Injectable()
-export class AseLinkInterceptorService extends LinkInterceptorService {
-
-  constructor(private _diagnosticServicePrivate: DiagnosticService, private _categoryServicePrivate: AseCategoryService) {
-    super(_diagnosticServicePrivate, _categoryServicePrivate);
-  }
-}
