--- AngularApp/projects/app-service-diagnostics/src/app/resources/web-sites/services/site-link-inspector.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { SiteLinkInspectorService } from './site-link-inspector.service';
-
-describe('SiteLinkInspectorService', () => {
-  let service: SiteLinkInspectorService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(SiteLinkInspectorService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
