+++ AngularApp/projects/app-service-diagnostics/src/app/resources/web-sites/web-sites.module.ts
@@ -21,9 +21,7 @@ import { QuickLinkService } from '../../shared-v2/services/quick-link.service';
 import { SiteQuickLinkService } from './services/site-quick-link.service';
 import { RiskAlertService } from '../../shared-v2/services/risk-alert.service';
 import { SiteRiskAlertService } from './services/site-risk-alert.service';
-import { DiagnosticDataModule, GenericLinkInterceptorService } from 'diagnostic-data';
-import { SiteLinkInspectorService } from './services/site-link-inspector.service';
-import { LinkInterceptorService } from '../../shared-v2/services/link-interceptor.service';
+import { DiagnosticDataModule} from 'diagnostic-data';
 
 const ResourceRoutes = RouterModule.forChild([
   {
@@ -65,14 +63,12 @@ const ResourceRoutes = RouterModule.forChild([
     ContentService,
     SiteFeatureService,
     SitesCategoryService,
-    SiteLinkInspectorService,
     LoggingV2Service,
     CXPChatCallerService,
     { provide: ResourceService, useExisting: WebSitesService },
     { provide: CategoryService, useClass: SitesCategoryService },
     { provide: FeatureService, useExisting: SiteFeatureService },
     { provide: SupportTopicService, useClass: SiteSupportTopicService },
-    { provide: LinkInterceptorService, useExisting: SiteLinkInspectorService },
     ResourceResolver,
     WebSiteFilter,
     { provide: QuickLinkService, useExisting: SiteQuickLinkService },
