--- AngularApp/projects/app-service-diagnostics/src/app/shared-v2/services/link-interceptor.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { LinkInterceptorService } from './link-interceptor.service';
-
-describe('LinkInterceptorService', () => {
-  let service: LinkInterceptorService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(LinkInterceptorService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
