+++ AngularApp/projects/diagnostic-data/src/lib/diagnostic-data.module.ts
@@ -169,7 +169,6 @@ import {OpenAIGenieComponent} from './components/openai-genie/openai-genie.compo
 import {OpenAIChatContainerComponent} from './components/openai-chat-container/openai-chat-container.component';
 
 import { FabDialogModule } from '@angular-react/fabric/lib/components/dialog';
-import { GenericLinkInterceptorService } from './services/generic-link-interceptor.service';
 
 @NgModule({
     imports: [
@@ -369,8 +368,7 @@ export class DiagnosticDataModule {
                 QueryResponseService,
                 OptInsightsGenericService,
                 DemoSubscriptions,
-                ChatUIContextService,
-                GenericLinkInterceptorService
+                ChatUIContextService
             ]
         };
     }
